\name{arrange_ggplot2}
\alias{arrange_ggplot2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
arrange_ggplot2
}
\description{
This function creates a grid on the graphics device and arranges ggplots on it in order.
}
\usage{
arrange_ggplot2(..., nrow = NULL, ncol = NULL, as.table = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{\dots}{
Ggplot2 plots for arrangement. Must be a list.
}
  \item{nrow}{
The number of rows in the grid
}
  \item{ncol}{
The number of columns in the grid.
}
  \item{as.table}{
If the grid is a table, set to TRUE.
}
}
