% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citv.R
\name{citv}
\alias{citv}
\title{Count data within intervals}
\usage{
citv(vec)
}
\arguments{
\item{vec}{A vector}
}
\description{
Count data within intervals
}
\examples{
citv(c("a","a","b","b","b"))
data(StraRoad)
citv(StraRoad$population)
}
