% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc.R
\name{disc}
\alias{disc}
\title{Convert continuous data to discretized data}
\usage{
disc(var, n, method = "quantile", ManualItv)
}
\arguments{
\item{var}{A numeric vector}

\item{n}{A number}

\item{method}{A character}

\item{ManualItv}{A numeric vector}
}
\description{
Convert continuous data to discretized data
}
\examples{
disc(rnorm(100,0,2), 6, method = "quantile")
disc(c(1:10,15:20,5:30,10:25), 5, method = "natural")
data(Roaddamage)
disc(Roaddamage$population, 6, method = "quantile")

}
