% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd.eco.R
\name{gd.eco}
\alias{gd.eco}
\title{Calculate ecological detectors}
\usage{
gd.eco(y, x)
}
\arguments{
\item{y}{A numeric vector}

\item{x}{A vector or a data.frame}
}
\description{
Calculate ecological detectors
}
\examples{
data(StraRoad)
gd.eco(StraRoad$damage, StraRoad[,c(2:5)])

}
