% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd.ecomatrix.R
\name{gd.ecomatrix}
\alias{gd.ecomatrix}
\title{Convert ecological detector results to ecological matrix}
\usage{
gd.ecomatrix(vec)
}
\arguments{
\item{vec}{A data.frame of ecological detector result of a strata variable}
}
\description{
Convert ecological detector results to ecological matrix
}
\examples{
data(StraRoad)
gdeco <- gd.eco(StraRoad$damage, StraRoad[,c(2:5)])
gd.ecomatrix(gdeco)

}
