% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd.interact.R
\name{gd.interact}
\alias{gd.interact}
\title{Calculate interaction detectors}
\usage{
gd.interact(y, x)
}
\arguments{
\item{y}{A numeric vector}

\item{x}{A data.frame with more than two strata variables}
}
\description{
Calculate interaction detectors
}
\examples{
data(StraRoad)
gd.riskmean(StraRoad$damage, StraRoad[,c(2:5)])

}
