% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd.R
\name{gd}
\alias{gd}
\alias{print.gd}
\alias{plot.gd}
\title{Calculate geographical detectors}
\usage{
gd(y, x)
\\method{print}{gd}(result)
\\method{plot}{gd}(result)
}
\arguments{
\item{y}{A numeric vector of response variable}

\item{x}{A vector or a data.frame of explanatory variables}

\item{result}{A list of factor detector results}
}
\description{
Calculate geographical detectors
}
\examples{
g1 <- gd(ndvi_40$NDVIchange, ndvi_40$Climatezone)
# g1

}
