% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd.risk.R
\name{gd.risk}
\alias{gd.risk}
\title{Calculate risk detectors}
\usage{
gd.risk(y, x)
}
\arguments{
\item{y}{A numeric vector}

\item{x}{A vector or a data.frame}
}
\description{
Calculate risk detectors
}
\examples{
gd.risk(c(1.2,1.5,2.1,3.3,3.6, 4.3,1.2,1.5,1.9,2.8),
        c("d","d","d","c","c", "c","d","d","d","c"))
data(StraRoad)
gd.risk(StraRoad$damage, StraRoad[,c(2:5)])

}
