% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd.riskmatrix.R
\name{gd.riskmatrix}
\alias{gd.riskmatrix}
\title{Convert risk detector results to risk matrix}
\usage{
gd.riskmatrix(vec)
}
\arguments{
\item{vec}{A data.frame of risk detector result of a strata variable}
}
\description{
Convert risk detector results to risk matrix
}
\examples{
data(StraRoad)
gdrisk <- gd.risk(StraRoad$damage, StraRoad[,c(2:5)])
gd.riskmatrix(gdrisk$population)

}
