% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd.riskmean.R
\name{gd.riskmean}
\alias{gd.riskmean}
\title{Calculate risk means within intervalss}
\usage{
gd.riskmean(y, x)
}
\arguments{
\item{y}{A numeric vector}

\item{x}{A vector or a data.frame}
}
\description{
Calculate risk means within intervalss
}
\examples{
gd.riskmean(c(1.2,1.5,2.1,3.3,3.6, 4.3,1.2,1.5,1.9,2.8),
            c("d","d","d","c","c", "c","d","d","d","c"))
data(StraRoad)
gd.riskmean(StraRoad$damage, StraRoad[,c(2:5)])

}
