% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdeco.R
\name{gdeco}
\alias{gdeco}
\alias{print.gdeco}
\alias{plot.gdeco}
\title{Calculate ecological detectors and ecological matrix and visulization}
\usage{
gdeco(y, x)
\\method{print}{gdeco}(result)
\\method{plot}{gdeco}(result)
}
\arguments{
\item{y}{A numeric vector of response variable}

\item{x}{A vector or a data.frame of explanatory variables}

\item{result}{A list of ecological detector results}
}
\description{
Calculate ecological detectors and ecological matrix and visulization
}
\examples{
ge1 <- gdeco(ndvi_40$NDVIchange, ndvi_40[,2:3])
# ge1

}
