% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdinteract.R
\name{gdinteract}
\alias{gdinteract}
\alias{print.gdinteract}
\alias{plot.gdinteract}
\title{Calculate interaction detectors and visualization. The types of interactions
include "Enhance, nonlinear", "Independent", "Enhance, bi-",
"Weaken, uni-" and "Weaken, nonlinear".}
\usage{
gdinteract(y, x)
\\method{print}{gdinteract}(result)
\\method{plot}{gdinteract}(result)
}
\arguments{
\item{y}{A numeric vector of response variable}

\item{x}{A vector or a data.frame of explanatory variables}

\item{result}{A list of interaction detector results}
}
\description{
Calculate interaction detectors and visualization. The types of interactions
include "Enhance, nonlinear", "Independent", "Enhance, bi-",
"Weaken, uni-" and "Weaken, nonlinear".
}
\examples{
gi1 <- gdinteract(ndvi_40$NDVIchange, ndvi_40[,2:3])
# gi1

}
