% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrisk.R
\name{gdrisk}
\alias{gdrisk}
\alias{print.gdrisk}
\alias{plot.gdrisk}
\title{Calculate risk detectors and risk matrix and visualization}
\usage{
gdrisk(y, x)
\\method{print}{gdrisk}(result)
\\method{plot}{gdrisk}(result)
}
\arguments{
\item{y}{A numeric vector of response variable}

\item{x}{A vector or a data.frame of explanatory variables}

\item{result}{A list of risk detector results}
}
\description{
Calculate risk detectors and risk matrix and visualization
}
\examples{
gr1 <- gdrisk(ndvi_40$NDVIchange, ndvi_40[,2:3])
# gr1
# plot(gr1)

}
