\name{liverpool_census}
\alias{liverpool_census}
\docType{data}
\title{
Liverpool_Census
}
\description{
Census Data for Liverpool from the 2001 Census. This should be used with the sample \code{\link{specification}} file also supplied with gd.
}
\usage{data(liverpool_census)}
\format{
  A data frame with 1501 rows (the areas in Liverpool) and 54 columns (various Census statistics).
}

\source{
UK Census data from 2001.
}


\keyword{datasets}
