% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd.R
\name{plot.gd}
\alias{plot.gd}
\title{plot factor detector results}
\usage{
\method{plot}{gd}(x, sig = TRUE, ...)
}
\arguments{
\item{x}{A list of factor detector results}

\item{sig}{If TRUE, only spatial associations that are significant at the 0.05 level will be plotted;
If FALSE, all spatial associations will be plotted.}

\item{...}{Ignore}
}
\description{
plot factor detector results
}
