% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd.R
\name{print.gd}
\alias{print.gd}
\title{print factor detector results}
\usage{
\method{print}{gd}(x, ...)
}
\arguments{
\item{x}{A list of factor detector results}

\item{...}{Ignore}
}
\description{
print factor detector results
}
