% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskmean.R
\name{riskmean}
\alias{riskmean}
\alias{print.riskmean}
\alias{plot.riskmean}
\title{Calculate risk means within intervals and visualization}
\usage{
riskmean(y, x)
\\method{print}{riskmean}(result)
\\method{plot}{riskmean}(result)
}
\arguments{
\item{y}{A numeric vector of response variable}

\item{x}{A vector or a data.frame of explanatory variables}

\item{result}{A data.frame of risk mean values}
}
\description{
Calculate risk means within intervals and visualization
}
\examples{
rm1 <- riskmean(ndvi_40$NDVIchange, ndvi_40[,2:3])
# rm1
# plot(rm1)

}
