\name{standardise}
\alias{standardise}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Standardise
}
\description{
Used by the \code{\link{cluster_types}} function to standardise the data.
}
\usage{
standardise(clusters, data_after_input)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{clusters}{
A vector of clusters.
}
  \item{data_after_input}{
The result of \code{\link{input_geodemographic}}
}
}
\value{
A standardised data frame.
}
\author{
Alex D. Singlegay, Liam Bratley
}
