% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stra.2v.R
\name{stra.2v}
\alias{stra.2v}
\title{Generate two category or discretized variables determined strata variables}
\usage{
stra.2v(var1, var2)
}
\arguments{
\item{var1}{A vector}

\item{var2}{A vector}
}
\description{
Generate two category or discretized variables determined strata variables
}
\examples{
stra.2v(c("a","a","b","b","b"), c("d","c","d","c","c"))
stra.2v(c(1,1,1,2,2, 3,1,2,2,3, 2,3,2,1,3), c(7,8,7,9,8, 7,8,9,8,9, 9,9,8,8,9))
data(StraRoad)
stra.2v(StraRoad$soiltype, StraRoad$population)

}
