% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stra2v.R
\name{stra2v}
\alias{stra2v}
\title{Generate two categorical or discretized variables determined strata variables}
\usage{
stra2v(categorical_var1, categorical_var2)
}
\arguments{
\item{categorical_var1}{A vector of a categorical variable}

\item{categorical_var2}{A vector of a categorical variable}
}
\description{
Generate two categorical or discretized variables determined strata variables
}
\examples{
stv2 <- stra2v(ndvi_40$Climatezone, ndvi_40$Mining)

}
