% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata2.R
\docType{data}
\name{StraRoad}
\alias{StraRoad}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\author{
Yongze Song \email{yongze.song@postgrad.curtin.edu.au}
}
\keyword{damage}
\keyword{dataset2}
\keyword{road}
