% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2m.R
\name{v2m}
\alias{v2m}
\title{Converts a vector to a lower triangular matrix.}
\usage{
v2m(vec, diag = FALSE)
}
\arguments{
\item{vec}{A data.frame of risk/interaction/ecological detector
result of a strata variable}

\item{diag}{TRUE/FALSE, indicating if the output matrix is a diagonal matrix.}
}
\description{
The function \code{v2m} is used in the functions
\code{gdrisk}, \code{gdinteract} and \code{gdeco} for converting
a vector is from the results of the risk detector result,
interaction detector result or ecological detector to a lower
triangular matrix.
}
