% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd.R
\name{gd}
\alias{gd}
\title{Calculate geographical detectors}
\usage{
gd(y, x)
}
\arguments{
\item{y}{A numeric vector}

\item{x}{A vector or a data.frame}
}
\description{
Calculate geographical detectors
}
\examples{
gd(c(1.2,1.5,2.1,3.3,3.6, 4.3,1.2,1.5,1.9,2.8),
   c("d","d","d","c","c", "c","d","d","d","c"))
data(StraRoad)
gd(StraRoad$damage, StraRoad[,c(2:5)])

}
