% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{H1N1}
\alias{H1N1}
\alias{h1n1_50}
\alias{h1n1_100}
\alias{h1n1_150}
\title{Spatial datasets of H1N1 flu incidences}
\format{A data frame with 3977 rows and 11 variables}
\usage{
h1n1_50

h1n1_100

h1n1_150
}
\description{
"H1N1" dataset is the provincial statistical incidences of
influenza A virus subtype H1N1 in China in 2013.
This dataset consists of H1N1 incidences and potential variables sampled from four sizes of
spatial grid units, including 10 km, 20 km, 50 km and 100 km.
The references of more details and data sources will be added.
}
\author{
Yongze Song \email{yongze.song@postgrad.curtin.edu.au}
}
\keyword{H1N1}
\keyword{dataset}
\keyword{datasets}
