% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NDVI}
\alias{NDVI}
\alias{ndvi_5}
\alias{ndvi_10}
\alias{ndvi_20}
\alias{ndvi_30}
\alias{ndvi_40}
\alias{ndvi_50}
\title{Spatial datasets of vegetation index changes}
\format{A data frame with 46295 rows and 7 variables}
\usage{
ndvi_5

ndvi_10

ndvi_20

ndvi_30

ndvi_40

ndvi_50
}
\description{
"NDVI" dataset is the NDVI change data from 2010 to 2014 in Inner Mongolia, China.
This dataset consists of NDVI change and potential variables sampled from six sizes of
spatial grid units, including 5 km, 10 km, 20 km, 30 km, 40 km and 50 km.
The references of more details and data sources will be added.
}
\author{
Yongze Song \email{yongze.song@postgrad.curtin.edu.au}
}
\keyword{NDVI}
\keyword{dataset}
\keyword{datasets}
