% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optidisc.R
\name{optidisc}
\alias{optidisc}
\title{Optimal discretization}
\usage{
optidisc(y = y, x = x, discmethod = discmethod, discitv = discitv)
}
\arguments{
\item{y}{A numeric vector of response variable}

\item{x}{A numeric vector of explanatory variable}

\item{discmethod}{A character vector of discretization methods}

\item{discitv}{A numeric vector of numbers of intervals}
}
\description{
Optimal discretization
}
\examples{
data(Roaddamage)
y <- Roaddamage$damage
x <- Roaddamage$population
discmethod <- c("equal","natural","quantile","geometric","sd")
discitv <- c(3:7)
optidisc(y, x, discmethod, discitv)

}
