% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesu.R
\name{sesu}
\alias{sesu}
\title{Comparison of size effects of spatial units.}
\usage{
sesu(gdlist, su)
}
\arguments{
\item{gdlist}{A list of \code{gdm} result or \code{gd} result}

\item{su}{A vector of sizes of spatial units}
}
\description{
Function for comparison of size effects of spatial units
in spatial heterogeneity analysis.
}
\examples{
ndvilist <- list(ndvi_30, ndvi_40, ndvi_50)
su <- c(30,40,50) ## sizes of spatial units
gdlist <- list() ## list of all geographical detectors results
## set optional parameters of optimal discretization
## optional methods: equal, natural, quantile, geometric, sd and manual
discmethod <- c("equal","natural","quantile")
discitv <- c(4:6)
## "gdm" function
for (i in 1:length(su)){
  ndvidata <- ndvilist[[i]]
  gdlist[[i]] <- gdm(NDVIchange ~ Climatezone + Mining + Tempchange + GDP,
                     continuous_variable = c("Tempchange", "GDP"),
                     data = ndvidata,
                     discmethod = discmethod, discitv = discitv)
}
sesu(gdlist, su) ## size effects of spatial units

}
